Moved CustomizationSection back to domain package
